Fix cgroup configuration format
From: Daniel Lezcano <dlezcano@fr.ibm.com>

This modification change the configuration format. Instead of creating
a 'cgroup' directory with a file per controller, a single file is used
to store the different value for the control groups. That allows to assign
several values to the same controller like "devices.allow" and keep the same
assignation order as defined in the configuration.

In order to keep compatibility, when the old cgroup format is detected, it
is automatically converted to the new format.

Signed-off-by: Daniel Lezcano <dlezcano@fr.ibm.com>