Change at compilation time the destruction of the network devices
From: Daniel Lezcano <dlezcano@fr.ibm.com>

The future kernel version will automatically autodestroy the network devices
when the network namespace exits. This is not the case for the current version.
In order to handle the both cases, I added a configuration option to disable
the network destruction when the container exits:
	--disable-network-destroy

Signed-off-by: Daniel Lezcano <dlezcano@fr.ibm.com>