List the available containers and the processes belonging to such container.
From: Daniel Lezcano <dlezcano@fr.ibm.com>

This modification change the lxc-ps command and adds the lxc-ls command.

The lxc-ps command takes the container name argument and shows the processes
belonging to the specified container. The usual ps argument can be passed to
the lxc-ps to change the output.
Examples:
  lxc-ps -n foo --forest
  lxc-ps -n foo -o pid=

The lxc-ls command list the container name available on the system. This is
useful to retrieve information for each container.
Examples:
  for i in $(lxc-ls); do
	lxc-info -n $i
	lxc-ps -n $i --forest
  done

Signed-off-by: Daniel Lezcano <dlezcano@fr.ibm.com>